네트워크 모듈 Package 변경
