+++ app/src/main/java/com/android/code/lib/koin/koinModules.kt
@@ -1,6 +1,7 @@
 package com.android.code.lib.koin
 
 import com.android.code.lib.network.MarvelService
+import com.android.code.lib.network.provideAPIClientService
 import com.android.code.repository.MarvelRepository
 import com.android.code.repository.MarvelRepositoryImpl
 import com.android.code.ui.main.MainViewModel
