+++ app/src/main/java/com/android/code/lib/network/NetworkModule.kt
@@ -1,5 +1,9 @@
-package com.android.code.lib.koin
+package com.android.code.lib.network
 
+import com.android.code.BuildConfig
+import com.android.code.util.addFlipperInterceptor
+import com.android.code.util.addStethoInterceptor
+import com.android.code.util.empty
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.google.gson.TypeAdapter
@@ -8,36 +12,14 @@ import com.google.gson.reflect.TypeToken
 import com.google.gson.stream.JsonReader
 import com.google.gson.stream.JsonToken
 import com.google.gson.stream.JsonWriter
-import com.android.code.BuildConfig
-import com.android.code.lib.network.APIClientInterceptor
-import com.android.code.util.addFlipperInterceptor
-import com.android.code.util.addStethoInterceptor
-import com.android.code.util.empty
+import okhttp3.Interceptor
 import okhttp3.OkHttpClient
+import okhttp3.Request
+import okhttp3.Response
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import java.io.IOException
-
-// API client interceptor
-private fun provideAPIClientInterceptor() = APIClientInterceptor()
-
-// Http logging interceptor
-private fun provideHttpLoggingInterceptor(): HttpLoggingInterceptor =
-    HttpLoggingInterceptor().apply { level = HttpLoggingInterceptor.Level.BODY }
-
-// Ok http client
-private fun provideOkHttpClient(
-    httpLoggingInterceptor: HttpLoggingInterceptor,
-    apiClientInterceptor: APIClientInterceptor,
-): OkHttpClient =
-    OkHttpClient.Builder()
-        .addInterceptor(httpLoggingInterceptor)
-        .addInterceptor(apiClientInterceptor)
-        .addStethoInterceptor()
-        .addFlipperInterceptor()
-        .build()
-
 
 // API client service
 internal inline fun <reified T : Any> provideAPIClientService(): T {
@@ -61,7 +43,47 @@ internal inline fun <reified T : Any> provideAPIClientService(): T {
         .create(T::class.java)
 }
 
-class StringNullToEmptyAdapterFactory : TypeAdapterFactory {
+private class ApiInterceptor : Interceptor {
+
+    @Throws(IOException::class)
+    override fun intercept(chain: Interceptor.Chain): Response =
+        response(chain.proceed(request(chain.request())))
+
+    private fun request(request: Request): Request =
+        request.newBuilder()
+            .addHeader("Content-Type", "application/json;charset=UTF-8")
+            .url(request.url)
+            .build()
+
+    private fun response(response: Response): Response {
+        return response
+            .newBuilder()
+            .addHeader("Content-Type", "application/json;charset=UTF-8")
+            .build()
+    }
+}
+
+// API client interceptor
+private fun provideAPIClientInterceptor() = ApiInterceptor()
+
+// Http logging interceptor
+private fun provideHttpLoggingInterceptor(): HttpLoggingInterceptor =
+    HttpLoggingInterceptor().apply { level = HttpLoggingInterceptor.Level.BODY }
+
+// Ok http client
+private fun provideOkHttpClient(
+    httpLoggingInterceptor: HttpLoggingInterceptor,
+    apiClientInterceptor: ApiInterceptor,
+): OkHttpClient =
+    OkHttpClient.Builder()
+        .addInterceptor(httpLoggingInterceptor)
+        .addInterceptor(apiClientInterceptor)
+        .addStethoInterceptor()
+        .addFlipperInterceptor()
+        .build()
+
+
+internal class StringNullToEmptyAdapterFactory : TypeAdapterFactory {
 
     @Suppress("UNCHECKED_CAST")
     override fun <T> create(gson: Gson?, type: TypeToken<T>): TypeAdapter<T>? {
@@ -74,7 +96,7 @@ class StringNullToEmptyAdapterFactory : TypeAdapterFactory {
     }
 }
 
-class StringNullToEmptyAdapter : TypeAdapter<String?>() {
+private class StringNullToEmptyAdapter : TypeAdapter<String?>() {
 
     @Throws(IOException::class)
     override fun read(reader: JsonReader): String {
